revenues/pyproject.toml
[tool.poetry]
authors = ["MuriloScarpaSitonio <muriloscarpa@gmail.com>"]
description = ""
name = "revenues"
version = "0.1.0"

[tool.poetry.dependencies]
fastapi = "^0.75.0"
fastapi-pagination = {extras = ["sqlalchemy"], version = "^0.9.1"}
pytest-mock = "^3.7.0"
python = "^3.8"
python-decouple = "^3.6"
sqlmodel = "^0.0.6"
uvicorn = {extras = ["standard"], version = "^0.17.6"}

[tool.poetry.dev-dependencies]
bandit = "^1.7.4"
black = "^22.1.0"
factory-boy = "^3.2.1"
mypy = "^0.942"
pytest = "7.1.1"
pytest-cov = "^3.0.0"
pytest-icdiff = "^0.5"
pytest-sugar = "^0.9.4"
pytest-xdist = "^2.5.0"
python-dateutil = "^2.8.2"
requests = "^2.27.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
line-length = 88

[tool.coverage.run]
omit = [
  "*tests/*",
]

[tool.pytest.ini_options]
addopts = "--disable-warnings"

[tool.coverage.report]
fail_under = 98
